MAINT: move cache helper functions to `sympy._cache`
